+++ lib/active_interaction/filters/symbol_filter.rb
@@ -17,10 +17,7 @@ module ActiveInteraction
     register :symbol
 
     def cast(value, _interaction)
-      case value
-      when Symbol
-        value
-      when String
+      if value.respond_to?(:to_sym)
         value.to_sym
       else
         super