+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - [#392][] - Integer parsing now defaults the base to 10. ([how to upgrade](#integer-parsing-base-now-10))
 - The `inputs` method now returns an `ActiveInteraction::Input` instead of a
   hash. The `ActiveInteraction::Input` still responds to all hash methods.
+- Implicit types are now supported for many filters:
+  - `string` accepts `to_str`
+  - `symbol` accepts `to_sym`
 
 ## Upgrading